requirements.txt
# 核心依赖
streamlit>=1.28.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
Pillow>=10.0.0
numpy>=1.24.0
edge-tts>=6.1.0
playsound>=1.3.0
pyttsx3>=2.90
requests>=2.31.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 文档生成
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
